% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_lambda}
\alias{calc_lambda}
\title{Calculate genomic inflation factor (lambda GC)}
\usage{
calc_lambda(p)
}
\arguments{
\item{p}{Numeric vector of p-values.}
}
\value{
Genomic inflation factor (lambda).
}
\description{
Calculate genomic inflation factor (lambda GC)
}
