% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-helpers.R
\name{filter_region}
\alias{filter_region}
\title{Filter variants by genomic region}
\usage{
filter_region(data, chr, start, end)
}
\arguments{
\item{data}{A \code{gwas_data} object or data.frame with CHR and BP columns.}

\item{chr}{Chromosome (integer or string like "chr6").}

\item{start}{Start position (bp).}

\item{end}{End position (bp).}
}
\value{
A filtered \code{gwas_data} object.
}
\description{
Subset a GWAS dataset to variants within a specified genomic region.
}
\examples{
data(example_gwas)
mhc <- filter_region(example_gwas, chr = 6, start = 25e6, end = 34e6)
}
