% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{gwas_palettes}
\alias{gwas_palettes}
\title{List available GWAS palettes}
\usage{
gwas_palettes()
}
\value{
A character vector of palette names.
}
\description{
List available GWAS palettes
}
\examples{
gwas_palettes()
}
