% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwasplot-package.R
\docType{package}
\name{gwasplot-package}
\alias{gwasplot}
\alias{gwasplot-package}
\title{gwasplot: Modern 'ggplot2' Visualizations for Genome-Wide Association Studies}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Create publication-ready Manhattan plots, QQ plots, Miami plots, locus zoom plots, and novel genome-wide heatmaps from GWAS summary statistics. Reads PLINK, REGENIE, GCTA, and GEMMA formats natively. All plot functions return 'ggplot2' objects for full customization. Smart downsampling handles datasets with 10 million+ variants efficiently.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bczech/gwasplot}
  \item Report bugs at \url{https://github.com/bczech/gwasplot/issues}
}

}
\author{
\strong{Maintainer}: Bartosz Czech \email{bartosz.w.czech@gmail.com} (\href{https://orcid.org/0000-0002-9908-3007}{ORCID})

Authors:
\itemize{
  \item Bartosz Czech \email{bartosz.w.czech@gmail.com} (\href{https://orcid.org/0000-0002-9908-3007}{ORCID})
}

}
\keyword{internal}
