% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-gcta.R
\name{read_gcta_mlma}
\alias{read_gcta_mlma}
\title{Read GCTA MLMA results}
\usage{
read_gcta_mlma(file, ...)
}
\arguments{
\item{file}{Path to a GCTA .mlma file.}

\item{...}{Additional arguments passed to \code{\link[data.table:fread]{data.table::fread()}}.}
}
\value{
A \code{gwas_data} object.
}
\description{
Read GCTA MLMA results
}
\examples{
f <- system.file("extdata", "example_gcta.mlma", package = "ggwas")
gwas <- read_gcta_mlma(f)
gwas
}
