% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-plink.R
\name{read_plink_assoc}
\alias{read_plink_assoc}
\title{Read PLINK association results}
\usage{
read_plink_assoc(file, ...)
}
\arguments{
\item{file}{Path to a PLINK .assoc file.}

\item{...}{Additional arguments passed to \code{\link[data.table:fread]{data.table::fread()}}.}
}
\value{
A \code{gwas_data} object.
}
\description{
Read PLINK association results
}
