% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-plink.R
\name{read_plink_logistic}
\alias{read_plink_logistic}
\title{Read PLINK logistic regression results}
\usage{
read_plink_logistic(file, test = "ADD", ...)
}
\arguments{
\item{file}{Path to a PLINK .assoc.logistic file.}

\item{test}{Which test to keep (default "ADD" for additive model).}

\item{...}{Additional arguments passed to \code{\link[data.table:fread]{data.table::fread()}}.}
}
\value{
A \code{gwas_data} object.
}
\description{
Read PLINK logistic regression results
}
\examples{
f <- system.file("extdata", "example_plink.assoc.logistic", package = "ggwas")
gwas <- read_plink_logistic(f)
gwas
}
