% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-regenie.R
\name{read_regenie}
\alias{read_regenie}
\title{Read REGENIE results}
\usage{
read_regenie(file, ...)
}
\arguments{
\item{file}{Path to a .regenie results file.}

\item{...}{Additional arguments passed to \code{\link[data.table:fread]{data.table::fread()}}.}
}
\value{
A \code{gwas_data} object.
}
\description{
Read REGENIE results
}
\examples{
f <- system.file("extdata", "example_regenie.regenie", package = "ggwas")
gwas <- read_regenie(f)
gwas
}
