% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-format.R
\name{validate_gwas_data}
\alias{validate_gwas_data}
\title{Validate a gwas_data object}
\usage{
validate_gwas_data(x)
}
\arguments{
\item{x}{A gwas_data object.}
}
\value{
Invisible x (raises warnings/errors on invalid data).
}
\description{
Validate a gwas_data object
}
