% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggwas-package.R
\docType{package}
\name{ggwas-package}
\alias{ggwas}
\alias{ggwas-package}
\title{ggwas: Modern 'ggplot2' Visualizations for Genome-Wide Association Studies}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Create publication-ready visualizations for genome-wide association studies. Provides 17 plot types including Manhattan, QQ, Miami, locus zoom, PheWAS, colocalization, fine-mapping, genetic correlation, SNP density, and density-signal comparison plots. Reads PLINK, REGENIE, GCTA, and GEMMA formats natively. All plot functions return 'ggplot2' objects for full customization. Smart downsampling handles datasets with 10 million+ variants efficiently.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bczech/ggwas}
  \item Report bugs at \url{https://github.com/bczech/ggwas/issues}
}

}
\author{
\strong{Maintainer}: Bartosz Czech \email{bartosz.w.czech@gmail.com} (\href{https://orcid.org/0000-0002-9908-3007}{ORCID})

Authors:
\itemize{
  \item Bartosz Czech \email{bartosz.w.czech@gmail.com} (\href{https://orcid.org/0000-0002-9908-3007}{ORCID})
}

}
\keyword{internal}
