% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-gemma.R
\name{read_gemma}
\alias{read_gemma}
\title{Read GEMMA association results}
\usage{
read_gemma(file, p_column = "p_wald", ...)
}
\arguments{
\item{file}{Path to a GEMMA .assoc.txt file.}

\item{p_column}{Which p-value column to use: "p_wald", "p_lrt", or "p_score".}

\item{...}{Additional arguments passed to \code{\link[data.table:fread]{data.table::fread()}}.}
}
\value{
A \code{gwas_data} object.
}
\description{
Read GEMMA association results
}
\examples{
\donttest{
gwas <- read_gemma("my_results.assoc.txt")
}
}
