% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-gwas.R
\name{scale_color_chromosome}
\alias{scale_color_chromosome}
\alias{scale_fill_chromosome}
\title{Chromosome color scale}
\usage{
scale_color_chromosome(colors = c("#2C3E50", "#7F8C8D"), ...)

scale_fill_chromosome(colors = c("#2C3E50", "#7F8C8D"), ...)
}
\arguments{
\item{colors}{Two-element character vector of alternating colors.}

\item{...}{Additional arguments passed to \code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.}
}
\value{
A ggplot2 color scale.
}
\description{
Alternating color scale for chromosomes in Manhattan plots.
}
