% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{scale_color_gwas}
\alias{scale_color_gwas}
\alias{scale_fill_gwas}
\title{Colorblind-safe chromosome color scale}
\usage{
scale_color_gwas(palette = "colorblind", ...)

scale_fill_gwas(palette = "colorblind", ...)
}
\arguments{
\item{palette}{Palette name from \code{\link[=gwas_palette]{gwas_palette()}}.}

\item{...}{Additional arguments passed to \code{\link[ggplot2:scale_color_manual]{ggplot2::scale_color_manual()}}.}
}
\value{
A ggplot2 color scale.
}
\description{
Scale using colorblind-friendly alternating colors for chromosomes.
}
